SimpleBankAccount : 4a : withdraw amount: create passing code

+++ src/test/java/edu/rit/swen352/tdd/SimpleBankAccountTest.java
@@ -63,8 +63,8 @@ class SimpleBankAccountTest {
     @DisplayName("Withdraw An Amount From The Account")
     void withdrawAmount() {
         SimpleBankAccount bankAccount = new SimpleBankAccount(100.99);
-        int remaining1 = bankAccount.withdraw(100);
-        int remaining2 = bankAccount.withdraw(bankAccount.getBalance());
+        double remaining1 = bankAccount.withdraw(100);
+        double remaining2 = bankAccount.withdraw(bankAccount.getBalance());
         assertAll("Withdraw Assertions",
                 () -> assertEquals(0.99, remaining1),
                 () -> assertEquals(0, remaining2));
